refactor: move card classes into a package card
