+++ src/main/java/edu/ntnu/idatt2003/millions/view/exchange/ExchangeOverview.java
@@ -3,10 +3,10 @@ package edu.ntnu.idatt2003.millions.view.exchange;
 import edu.ntnu.idatt2003.millions.manager.GameManager;
 import edu.ntnu.idatt2003.millions.model.exchange.Exchange;
 import edu.ntnu.idatt2003.millions.observer.GameObserver;
-import edu.ntnu.idatt2003.millions.view.exchange.overview.GainersCard;
-import edu.ntnu.idatt2003.millions.view.exchange.overview.LosersCard;
+import edu.ntnu.idatt2003.millions.view.exchange.overview.card.GainersCard;
+import edu.ntnu.idatt2003.millions.view.exchange.overview.card.LosersCard;
 import edu.ntnu.idatt2003.millions.view.exchange.overview.StockRankingTable;
-import edu.ntnu.idatt2003.millions.view.exchange.overview.TotalStocksCard;
+import edu.ntnu.idatt2003.millions.view.exchange.overview.card.TotalStocksCard;
 import javafx.scene.layout.ColumnConstraints;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.HBox;
